Added option "PICO_TCP_NODELAY" to pico_socket_setoption.
A check is
made first to make sure the socket is using TCP.
To store the socket options, the "opt_flags" field is added to the
pico_socket struct.
By default when opening a TCP socket, the Nagle option is enabled. So
the user can disable this feature with the TCP_NODELAY option.

To get the status of the option, the pico_socket_getoption returns
the current setting of the TCP_NODELAY flag.