Task #267 - completed TCP check
Checked tcp implementation of closing with sending events to
socket. Changed picoapp behaviour of tcpclient and tcpecho.
Tcpecho now waits on all data to come in from tcpclient, so
now all data is echoed. Tcpclient will only send data on the
EV_WR call, this reduces the calling frequency of the
socket_write function.

The tests now don't show anymore retransmissions and the whole
data buffer is sent between 2 picoapps.

Note: data loss is possible on the tcpecho side since it discards
data in the callback when it can't write to the socket (full).