Updated timers unittest
When inserting a timer with a lower expire time than an already
existing timer, the sequence of the timers is rearranged.

With the timer for ARP flooding protection being already present, and
having a higher expire time than the timers that were added in the test,
the order in which the timers are stored was not linear anymore. This
caused the unittest to fail as it expects the timers to be stored
in the same order they were inserted.

The expire time for the unittest timers is set to a very large
number to avoid this problem.