TCP: Adopt sent frame size to smaller RCV window.
If the advertised window is smaller than the next segment in the output queue,
TCP will split the current segment into two smaller segments, and send exactly
the amount of bytes advertised. The remaining bytes are enqueued once more in
the output queue for later send.

Should fix #127, traces are positive, but the "window shrinking" test still fails.