IPv4: Check packet len before processing
When a packet is received, the length in the header must be checked
against the actual IP buffer length. If the header length has been
altered, or it's set to a bigger value on purpose by an attacker, the
CRC function may violate the heap memory boundaries.

Discovered using Pareon Verify.