Remove optional options from Option Request option
In 17.1.1 from RFC3315:

"The client SHOULD include an Option Request option (see section 22.7)
to indicate the options the client is interested in receiving. The
client MAY additionally include instances of those options that are
identified in the Option Request option."

If there are no options, then the length of the option is zero.
Later we should add the possibility to provide options, but for now
this is ok.